+++ app/src/main/java/com/example/myapplication/Bottle.kt
@@ -2,4 +2,4 @@ package com.example.myapplication
 
 import java.io.Serializable
 
-data class Bottle(val name: String, val price: Int): Serializable;
+data class Bottle(val name: String, val price: Double): Serializable;
